+++ modules/util/portal-tools-sample-sql-builder/src/main/java/com/liferay/portal/tools/sample/sql/builder/SampleSQLBuilder.java
@@ -377,6 +377,8 @@ public class SampleSQLBuilder {
 		).put(
 			"socialActivityDataFactory", new SocialActivityDataFactory()
 		).put(
+			"subscriptionDataFactory", new SubscriptionDataFactory()
+		).put(
 			"userDataFactory", new UserDataFactory()
 		).put(
 			"wikiDataFactory", new WikiDataFactory()
