+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/mb.ftl
@@ -11,7 +11,7 @@
 	<#list mbThreadModels as mbThreadModel>
 		${dataFactory.toInsertSQL(mbThreadModel)}
 
-		${dataFactory.toInsertSQL(dataFactory.newSubscriptionModel(mbThreadModel))}
+		${dataFactory.toInsertSQL(subscriptionDataFactory.newSubscriptionModel(mbThreadModel))}
 
 		<@insertAssetEntry _entry=mbThreadModel />
